+++ WolvenKit.App/Helpers/WKitUIScripting.cs
@@ -5,7 +5,6 @@ using System.IO;
 using System.Linq;
 using System.Reactive;
 using Microsoft.ClearScript;
-using SharpDX;
 using WolvenKit.App.Factories;
 using WolvenKit.App.Services;
 using WolvenKit.Common;
@@ -61,8 +60,8 @@ public class WKitUIScripting : WKitScripting
     /// <summary>
     /// Add the specified cr2w file to the project from the game archives.
     /// </summary>
-    /// <param name="path"></param>
-    /// <param name="cr2w"> File to be saved</param>
+    /// <param name="path">The file to write to</param>
+    /// <param name="cr2w">File to be saved</param>
     public virtual void SaveToProject(string path, CR2WFile cr2w)
     {
         if (_projectManager.ActiveProject is null)
@@ -80,8 +79,8 @@ public class WKitUIScripting : WKitScripting
     /// <summary>
     /// Add the specified gameFile file to the project from the game archives.
     /// </summary>
-    /// <param name="path"></param>
-    /// <param name="gameFile"> File to be saved</param>
+    /// <param name="path">The file to write to</param>
+    /// <param name="gameFile">File to be saved</param>
     public virtual void SaveToProject(string path, IGameFile gameFile)
     {
         if (_projectManager.ActiveProject is null)
@@ -98,8 +97,8 @@ public class WKitUIScripting : WKitScripting
     /// <summary>
     /// Save the specified file to the project raw folders, in either json or CR2W
     /// </summary>
-    /// <param name="path"></param>
-    /// <param name="content"></param>
+    /// <param name="path">The file to write to</param>
+    /// <param name="content">The string to write to the file</param>
     public virtual void SaveToRaw(string path, string content)
     {
         if (_projectManager.ActiveProject is null)
@@ -136,8 +135,8 @@ public class WKitUIScripting : WKitScripting
     /// <summary>
     /// Loads the specified game file from the project files rather than game archives.
     /// </summary>
-    /// <param name="path"></param>
-    /// <param name="type"></param>
+    /// <param name="path">The file to open for reading</param>
+    /// <param name="type">The type of the object which is returned. Can be "cr2w" or "json"</param>
     /// <returns></returns>
     public virtual object? LoadGameFileFromProject(string path, string type)
     {
@@ -183,8 +182,8 @@ public class WKitUIScripting : WKitScripting
     /// <summary>
     /// Loads the specified json file from the project raw files rather than game archives.
     /// </summary>
-    /// <param name="path"></param>
-    /// <param name="type"></param>
+    /// <param name="path">The file to open for reading</param>
+    /// <param name="type">The type of the object which is returned. Can be "cr2w" or "json"</param>
     /// <returns></returns>
     public virtual object? LoadRawJsonFromProject(string path, string type)
     {
@@ -479,6 +478,13 @@ public class WKitUIScripting : WKitScripting
             expVM.ProcessSelectedCommand.Execute(Unit.Default);
         }
     }
+
+    /// <summary>
+    /// Check if file exists in either the game archives or the project
+    /// </summary>
+    /// <param name="path">file path to check</param>
+    /// <returns></returns>
+    public override bool FileExists(string path) => base.FileExists(path);
 
     /// <summary>
     /// Check if file exists in either the game archives or the project
