+++ WolvenKit.Modkit/Scripting/WKitScripting.cs
@@ -33,7 +33,7 @@ public class WKitScripting
     /// <summary>
     /// Loads a file from the base archives using either a file path or hash
     /// </summary>
-    /// <param name="path"></param>
+    /// <param name="path">The path of the file to retrieve</param>
     /// <returns></returns>
     [Description("GetFileFromBase")]
     public virtual IGameFile? GetFileFromBase(string path)
@@ -65,7 +65,7 @@ public class WKitScripting
     /// <summary>
     /// Loads a file from the base archives using either a file path or hash
     /// </summary>
-    /// <param name="hash"></param>
+    /// <param name="hash">The hash of the file to retrieve</param>
     /// <returns></returns>
     public virtual IGameFile? GetFileFromBase(ulong hash)
     {
@@ -80,7 +80,7 @@ public class WKitScripting
     /// <summary>
     /// Creates a json representation of the specifed game file.
     /// </summary>
-    /// <param name="gameFile"></param>
+    /// <param name="gameFile">The gameFile which should be converted</param>
     /// <returns></returns>
     /// <exception cref="ArgumentNullException"></exception>
     [Description("GameFileToJson")]
@@ -120,7 +120,7 @@ public class WKitScripting
     public virtual string ChangeExtension(string path, string extension) => Path.ChangeExtension(path, extension);
 
     /// <summary>
-    /// Check if file exists in either the game archives or the project
+    /// Check if file exists in the game archives
     /// </summary>
     /// <param name="path">file path to check</param>
     /// <returns></returns>
@@ -135,9 +135,9 @@ public class WKitScripting
     }
 
     /// <summary>
-    /// 
+    /// Check if file exists in the game archives
     /// </summary>
-    /// <param name="hash"></param>
+    /// <param name="hash">hash value to be checked</param>
     /// <returns></returns>
     public virtual bool FileExists(ulong hash)
     {
